+++ src/controllers/SettingsController.php
@@ -28,7 +28,8 @@ class SettingsController extends Controller
 			'fieldSets' => Plugin::getInstance()->getFieldSets()->getAllFieldSets(),
 			'settings' => $settings,
 			'displayTypeOptions' => DisplayType::options(DisplayType::cases()),
-			'defaultDisplayTypeOptions' => DisplayType::options($settings->getAvailableDisplayTypes($settings->defaultDisplayType)),
+			'defaultDisplayTypeOptions' => DisplayType::options($settings->getAvailableDisplayTypes()),
+			'defaultDisplayType' => $settings->getDefaultDisplayType()->value,
 			'productTypeOptions' => array_map(
 				static fn (ProductType $productType): array => [
 					'label' => $productType->name,